move storage+supplier from inventory to master
