+++ index.js
@@ -1,5 +1,5 @@
 var inventoryMap = new Map();
-inventoryMap.set("storage", require('./src/managers/inventory/storage-manager'));
+inventoryMap.set("transfer-in-doc-ext", require('./src/managers/inventory/transfer-in-doc-ext-manager'));
 inventoryMap.set("transfer-in-doc", require('./src/managers/inventory/transfer-in-doc-manager'));
 inventoryMap.set("transfer-out-doc", require('./src/managers/inventory/transfer-out-doc-manager'));
 inventoryMap.set("inventory", require('./src/managers/inventory/inventory-manager'));
@@ -21,9 +21,7 @@ inventoryMap.set("efr-tb-brd", require('./src/managers/inventory/efr-tb-brd-mana
 inventoryMap.set("efr-tb-brt", require('./src/managers/inventory/efr-tb-brt-manager'));
 inventoryMap.set("efr-tb-sab", require('./src/managers/inventory/efr-tb-sab-manager'));
 inventoryMap.set("efr-tb-alt", require('./src/managers/inventory/efr-tb-alt-manager'));
-inventoryMap.set("efr-hp-fng", require('./src/managers/inventory/efr-hp-fng-manager'));
-inventoryMap.set("supplier", require('./src/managers/inventory/supplier-manager'));
-inventoryMap.set("transfer-in-doc-ext", require('./src/managers/inventory/transfer-in-doc-ext-manager'));
+inventoryMap.set("efr-hp-fng", require('./src/managers/inventory/efr-hp-fng-manager')); 
 
 var merchandiserMap = new Map();
 merchandiserMap.set("efr-pk", require('./src/managers/merchandiser/efr-pk-manager'));
@@ -36,9 +34,7 @@ module.exports = {
         AccountManager: require("./src/managers/auth/account-manager"),
         RoleManager: require("./src/managers/auth/role-manager")
     },
-    inventory: {
-        StorageManager: inventoryMap.get("storage"),
-        StoreManager: inventoryMap.get("store"),
+    inventory: { 
         TransferInDocManager: inventoryMap.get("transfer-in-doc"),
         TransferOutDocManager: inventoryMap.get("transfer-out-doc"),
         InventoryManager: inventoryMap.get("inventory"),
@@ -60,7 +56,6 @@ module.exports = {
         PusatReturTokoTerimaBarangReturSelesaiPerbaikanManager: inventoryMap.get("efr-tb-brd"),
         PusatReturTokoTerimaBarangReturManager: inventoryMap.get("efr-tb-brt"),
         FinishingTerimaKomponenManager: inventoryMap.get("efr-tb-sab"),
-        SupplierManager: inventoryMap.get("supplier"),
         TransferInDocExtManager: inventoryMap.get("transfer-in-doc-ext"),
         map: inventoryMap
     }, 
@@ -77,6 +72,8 @@ module.exports = {
         CardTypeManager: require('./src/managers/master/card-type-manager'),
         StoreManager: require('./src/managers/master/store-manager'),
         ItemManager: require('./src/managers/master/item-manager'),
+        SupplierManager: require('./src/managers/master/supplier-manager'), 
+        StorageManager: require('./src/managers/master/storage-manager'),
         FinishedGoodsManager: require('./src/managers/master/finished-goods-manager')
     },
     sales: {
