+++ src/managers/inventory/efr-hp-fng-manager.js
@@ -27,7 +27,7 @@ module.exports = class FinishedGoodsManager {
         this.transferInDocCollection = this.db.use(map.inventory.TransferInDoc);
         this.transferOutDocCollection = this.db.use(map.inventory.TransferOutDoc);
 
-        var StorageManager = require('./storage-manager');
+        var StorageManager = require('../master/storage-manager');
         this.storageManager = new StorageManager(db, user);
 
         var ItemManager = require('../master/item-manager');
