+++ src/managers/inventory/efr-kb-rtt-manager.js
@@ -27,7 +27,7 @@ module.exports = class TokoTransferStokManager {
         this.expeditionDocCollection = this.db.use(map.inventory.ExpeditionDoc);
         this.SPKDocCollection = this.db.use(map.merchandiser.SPKDoc);
 
-        var StorageManager = require('./storage-manager');
+        var StorageManager = require('../master/storage-manager');
         this.storageManager = new StorageManager(db, user);
 
         var ItemManager = require('../master/item-manager');
