+++ src/managers/inventory/efr-tb-sab-manager.js
@@ -19,7 +19,7 @@ module.exports = class FinishingTerimaKomponenManager {
         this.db = db;
         this.user = user;
         this.transferInDocCollection = this.db.use(map.inventory.TransferInDoc);
-        var StorageManager = require('./storage-manager');
+        var StorageManager = require('../master/storage-manager');
         this.storageManager = new StorageManager(db, user);
 
         var ItemManager = require('../master/item-manager');
