+++ test/managers/storage-manager-test.js
@@ -1,10 +1,10 @@
 var should = require('should');
 var helper = require('../helper');
-var validate = require('bateeq-models').validator.inventory;
+var validate = require('bateeq-models').validator.master;
 var manager;
 
 function getData() {
-    var Storage = require('bateeq-models').inventory.Storage;
+    var Storage = require('bateeq-models').master.Storage;
     var storage = new Storage();
 
     var now = new Date();
@@ -23,7 +23,7 @@ function getData() {
 before('#00. connect db', function(done) {
     helper.getDb()
         .then(db => {
-            var StorageManager = require('../../src/managers/inventory/storage-manager');
+            var StorageManager = require('../../src/managers/master/storage-manager');
             manager = new StorageManager(db, {
                 username: 'unit-test'
             });
